Coalesce mdlManager slot clears

+++ src/Event/Fcl/fclShopMisc.c
@@ -293,7 +293,7 @@ void FUN_003f7890(int param_1,int param_2,u8 param_3,int param_4,int param_5,int
 void FUN_003f7a80(u64 param_1,u64 param_2,u32 param_3,int param_4,int param_5);
 void FUN_003f7c60(u32 param_1,u32 param_2,u32 param_3,int param_4,int param_5);
 void FUN_003f7d50(int param_1,int param_2,u64 param_3,int param_4, int param_5);
-void FUN_003f7fe0(u64 param_1,u64 param_2,u32 param_3,int param_4,int param_5 );
+void FUN_003f7fe0(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f8180(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
 void FUN_003f85a0(int param_1,int param_2,u64 param_3,int param_4, int param_5);
 void FUN_003f86a0(int param_1,int param_2,u32 param_3,int param_4,int param_5 );
@@ -6730,7 +6730,7 @@ void FUN_003f7d50(int param_1,int param_2,u64 param_3,int param_4,
 // FUN_003F7FE0 NONMATCHING
 
 
-void FUN_003f7fe0(u64 param_1,u64 param_2,u32 param_3,int param_4,int param_5
+void FUN_003f7fe0(int param_1,int param_2,u32 param_3,int param_4,int param_5
 
                  )
